tests/integration/test_end_to_end.py
"""Integration tests for end-to-end document processing."""
import pytest
from pathlib import Path
from PIL import Image
import numpy as np

# These would import actual modules in production
# from inference.olmocr_adapter import OlmOCRAdapter
# from preprocessing.layout_detector import CVLayoutDetector
# from preprocessing.image_enhancer import ImageEnhancer
# from validation.rule_engine import ValidationEngine
# from validation.schema_mapper import SchemaMapper


@pytest.fixture
def sample_document():
    """Create sample document image."""
    img = Image.new('RGB', (2550, 3300), color='white')
    return img


def test_full_extraction_pipeline(sample_document, tmp_path):
    """Test complete extraction pipeline."""
    # This is a scaffold for full integration test
    # In production, would test:
    # 1. Image enhancement
    # 2. Layout detection
    # 3. Region extraction
    # 4. olmOCR inference
    # 5. Self-consistency
    # 6. Validation
    # 7. Schema mapping
    
    # For now, basic structure test
    assert sample_document is not None
    assert sample_document.size == (2550, 3300)


def test_confidence_scoring_integration():
    """Test confidence scoring with validation."""
    # Mock prediction
    prediction = {
        "patient_id": "12345",
        "date": "10/30/2025",
        "checkboxes": [
            {"label": "Fever", "checked": True}
        ]
    }
    
    metadata = {
        "self_consistency_agreement": 1.0,
        "validation_pass_rate": 1.0,
        "difficulty_score": 0.3,
        "region_type": "checkbox",
        "prompt_id": "checkbox_v1"
    }
    
    # Compute confidence (scaffold)
    confidence = (
        0.40 * metadata["self_consistency_agreement"] +
        0.30 * metadata["validation_pass_rate"] +
        0.20 * (1 - metadata["difficulty_score"]) +
        0.10 * 0.85  # Historical accuracy
    )
    
    assert 0.5 <= confidence <= 1.0


def test_active_learning_feedback_loop():
    """Test active learning sample selection."""
    # Mock predictions with varying confidence
    predictions = [
        {"document_id": f"doc_{i}", "region_id": f"reg_{i}", 
         "confidence": np.random.uniform(0.5, 0.95),
         "difficulty": np.random.uniform(0.3, 0.8),
         "region_type": "checkbox"}
        for i in range(20)
    ]
    
    # Select low confidence samples
    low_conf = [p for p in predictions if p["confidence"] < 0.8]
    
    assert len(low_conf) >= 0
    assert all(p["confidence"] < 0.8 for p in low_conf)


def test_model_routing():
    """Test model router selection."""
    routing_config = {
        "strategy": "weighted",
        "models": {
            "olmocr-v1": {"weight": 0.8},
            "olmocr-v2": {"weight": 0.2}
        }
    }
    
    # Simulate routing
    import random
    random.seed(42)
    
    models = list(routing_config["models"].keys())
    weights = [routing_config["models"][m]["weight"] for m in models]
    
    selected = random.choices(models, weights=weights)[0]
    
    assert selected in models

