pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raelm"
version = "0.1.0"
description = "RaeLM: Reliable Agentic Extraction with Language Models"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "RaeLM Team"}
]

dependencies = [
    "mlflow>=2.9.0",
    "dvc[s3]>=3.30.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "celery[redis]>=5.3.4",
    "redis>=5.0.1",
    "opencv-python>=4.8.1",
    "pymupdf>=1.23.8",
    "pillow>=10.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.4",
    "rapidfuzz>=3.5.2",
    "streamlit>=1.28.2",
    "plotly>=5.17.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "psycopg2-binary>=2.9.9",
    "boto3>=1.29.7",
    "reportlab>=4.0.8",
    "pdf2image>=1.16.3",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "mlx-vlm>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
    "black>=23.11.0",
    "memory-profiler>=0.61.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools]
packages = ["ml_platform", "inference", "preprocessing", "validation", "api", "ui"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ml_platform --cov=inference --cov=preprocessing --cov=validation --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["ml_platform", "inference", "preprocessing", "validation", "api"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

